test file for embedding project
This page includes a Snap! window embedded inside a FYTD.  (The project
isn't relevant to the page -- it was just an example.)  I don't know
how to make this work with xml files in the curriculum unless we
hardcode the location at bjc.edc.org into the page.  But we could make
a BJC account like the BJCSolutions account, and save starter projects
there.