U1L2 vocab and bjc.css...
... to add "vocabFullWidth" class.

Also on page 4, commented out the claim that this is Machine Learning;
there's a sense in which it's true, but I think it's very misleading,
especially the part about cutting-edge, which is about /not/ burning
the algorithm into the program.  I could be talked out of this if
someone feels strongly about it.