1.3 temporary changes, see below
Cindy want to data-gifffer the animated gifs in 1.3 for her class
tomorrow, but the play button is invisible against the black
background.  The long term solution is to invert the video so it's
black on yellow instead of yellow on black, but for now I put an ugly
"click me" next to two pictures, on 1.3.3 and 1.3.4.  Will fix this
over the weekend.