add wordlists to bjc-r/data
New directory made for these.  Unsorted and sorted versions.  "words"
is about 200K words and takes no time, even to sort (using the block in
the "bigger data" library).  "scientific" is almost 700K words.  It is
fast to filter (find 5-letter words, etc) in Snap! 5.0, blink and it's
done, but sorting it still takes a long time, so give them the sorted
version.

Importing in 5.0 takes no time, so I didn't make a starter project; we
can do that when we get the actual page settled.