New higher quality word lists
The ones we paid €280 for, filtering out cuss words and non-ASCII
letters except allowing apostrophe.  Every list is exactly the right
number of words.  Zebra isn't quite the last word in the big ones.  (I
had to cheat to get it in any but the 100K list.)  This commit also
includes the project containing those lists preloaded.