4.4.6 three EKs to go
Two will be easy after some sleep, but I cannot for the life of me
understand what "An understanding of the purpose of a computing
innovation provides developers with an improved ability to develop that
computing innovation" is supposed to mean.  It's easier to write a
program if you decide what program you want to write?