CSS - update for you to see
Hi Brian, some of these changes may break things for you (such as leaving weird non-working characters on the solutions pages), but I wanted you to see and have some time to react to the design change idea I had yesterday. In short: anything on a white background is required.