2.5.2: fix free-ish license desriptions
I made the bullets correct but terse, and added read-more yellow boxes to explain more about the differences between free and open software.  I'm not sure why we have "open access" in there at all; the typical high school student doesn't know anything about the economics of scholarly publication.  Is it in the Framework?

I also changed the hyphen in "for-profit" in the yellow box just above that gray box to a space.  People like us who work at nonprofits use "for-profit" as a compound adjective to describe, you know, the enemy, but regular people don't talk that way, and in this context it's a regular old prepositional phrase, not modifying a noun.

I also argued back in the red box you put around the yellow box in 1.4.1 that introduces the read-more symbol.  I think we already had this conversation the last time you wanted to remove that yellow box, although I could be wrong.